# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Contact Importer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUrl {
  url: string;
}

export interface AuthUser {
  name: string;
  email: string;
  /** @nullable */
  picture: string | null;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export interface ContactTextInput {
  /** Unstructured contact text to parse */
  text: string;
}

export interface ParsedContact {
  givenName: string;
  familyName: string;
  /**
     * Alternative name in another language
     * @nullable
     */
  nameAlternative?: string | null;
  /** @nullable */
  company?: string | null;
  /**
     * Alternative company name in another language
     * @nullable
     */
  companyAlternative?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  emails?: string[];
  phones?: string[];
  websites?: string[];
  /** @nullable */
  notes?: string | null;
}

export interface ContactUpdateInput {
  /** Google People API resource name (e.g. people/c1234) */
  resourceName: string;
  /** Free-text correction message from user */
  correction: string;
  currentContact: ParsedContact;
}

export interface CreatedContact {
  resourceName: string;
  contact: ParsedContact;
}

export type GoogleAuthCallbackParams = {
code: string;
state?: string;
};