# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Contact Importer API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Google OAuth authentication
  - name: contacts
    description: Contact parsing and creation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/google:
    get:
      operationId: googleAuthStart
      tags: [auth]
      summary: Start Google OAuth flow
      description: Returns the Google OAuth redirect URL
      responses:
        "200":
          description: OAuth URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUrl"

  /auth/google/callback:
    get:
      operationId: googleAuthCallback
      tags: [auth]
      summary: Google OAuth callback
      description: Exchanges code for tokens and stores in session
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to app after auth

  /auth/me:
    get:
      operationId: getAuthMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Authenticated user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Clear session and log out
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /contacts/parse:
    post:
      operationId: parseContact
      tags: [contacts]
      summary: Parse unstructured text into a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactTextInput"
      responses:
        "200":
          description: Parsed contact data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParsedContact"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contacts/create:
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact in Google Contacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParsedContact"
      responses:
        "200":
          description: Created contact confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedContact"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /contacts/update:
    post:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact via chat correction message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdateInput"
      responses:
        "200":
          description: Updated contact confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedContact"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    AuthUrl:
      type: object
      required: [url]
      properties:
        url:
          type: string

    AuthUser:
      type: object
      required: [name, email, picture]
      properties:
        name:
          type: string
        email:
          type: string
        picture:
          type: ["string", "null"]

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    ContactTextInput:
      type: object
      required: [text]
      properties:
        text:
          type: string
          description: Unstructured contact text to parse

    ParsedContact:
      type: object
      required: [givenName, familyName]
      properties:
        givenName:
          type: string
        familyName:
          type: string
        nameAlternative:
          type: ["string", "null"]
          description: Alternative name in another language
        company:
          type: ["string", "null"]
        companyAlternative:
          type: ["string", "null"]
          description: Alternative company name in another language
        jobTitle:
          type: ["string", "null"]
        emails:
          type: array
          items:
            type: string
        phones:
          type: array
          items:
            type: string
        websites:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]

    ContactUpdateInput:
      type: object
      required: [resourceName, correction, currentContact]
      properties:
        resourceName:
          type: string
          description: Google People API resource name (e.g. people/c1234)
        correction:
          type: string
          description: Free-text correction message from user
        currentContact:
          $ref: "#/components/schemas/ParsedContact"

    CreatedContact:
      type: object
      required: [resourceName, contact]
      properties:
        resourceName:
          type: string
        contact:
          $ref: "#/components/schemas/ParsedContact"