# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Contact Importer API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  name: string;
  email: string;
  /** @nullable */
  picture: string | null;
}