# lib/api-zod/src/generated/types/contactTextInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Contact Importer API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactTextInput {
  /** Unstructured contact text to parse */
  text: string;
}