# lib/api-zod/src/generated/types/contactUpdateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Contact Importer API
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedContact } from './parsedContact';

export interface ContactUpdateInput {
  /** Google People API resource name (e.g. people/c1234) */
  resourceName: string;
  /** Free-text correction message from user */
  correction: string;
  currentContact: ParsedContact;
}