# lib/api-zod/src/generated/types/createdContact.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Contact Importer API
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedContact } from './parsedContact';

export interface CreatedContact {
  resourceName: string;
  contact: ParsedContact;
}