# lib/api-zod/src/generated/types/parsedContact.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Contact Importer API
 * OpenAPI spec version: 0.1.0
 */

export interface ParsedContact {
  givenName: string;
  familyName: string;
  /**
     * Alternative name in another language
     * @nullable
     */
  nameAlternative?: string | null;
  /** @nullable */
  company?: string | null;
  /**
     * Alternative company name in another language
     * @nullable
     */
  companyAlternative?: string | null;
  /** @nullable */
  jobTitle?: string | null;
  emails?: string[];
  phones?: string[];
  websites?: string[];
  /** @nullable */
  notes?: string | null;
}